# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Booklab - backend",
        "description": "Booklab project API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/cardapio-items": {
            "get": {
                "tags": [
                    "CardapioItems"
                ],
                "summary": "Listagem de CardapioItems",
                "description": "Get CardapiItems",
                "operationId": "c2b73cd56a6dce30004b285dde4a9469",
                "responses": {
                    "default": {
                        "description": "List Items"
                    }
                }
            },
            "post": {
                "tags": [
                    "CardapioItems"
                ],
                "summary": "Add new",
                "description": "Relacionar um item a um cardapio",
                "operationId": "50a7599181b6a15ec4c59d4a8442373d",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "item_id": {
                                        "type": "integer"
                                    },
                                    "cardapio_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "item_id": 1,
                                    "cardapio_id": 1
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/cardapio-items/{id}": {
            "get": {
                "tags": [
                    "CardapioItems"
                ],
                "summary": "show",
                "description": "Show CardapioItems",
                "operationId": "7e983ce772108568ef42cb6727cc2045",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Parameter ",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "examples": {
                            "int": {
                                "summary": "An int value.",
                                "value": "1"
                            }
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": "OK"
                    }
                }
            },
            "put": {
                "tags": [
                    "CardapioItems"
                ],
                "summary": "Updates a CardapioItems",
                "operationId": "cc4bc234725fabe8e02ab8a4f4f307b8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Update",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "examples": {
                            "int": {
                                "summary": "item_id",
                                "value": "1"
                            }
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "item_id": {
                                        "type": "integer"
                                    },
                                    "cardapio_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "item_id": 1,
                                    "cardapio_id": 1
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "CardapioItems"
                ],
                "summary": "Delete CardapioItems",
                "description": "Delete CardapioItems",
                "operationId": "113526bf5596732f7f3e2a5901f8ee57",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Parameter ",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "examples": {
                            "int": {
                                "summary": "An int value.",
                                "value": "1"
                            }
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/categorias": {
            "post": {
                "tags": [
                    "Categorias"
                ],
                "summary": "Add new Categoria",
                "operationId": "7bf305bf13c77a93af29f16d53e3d6de",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nome": {
                                        "type": "string"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "nome": "Saída"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/categorias/{id}?with[]=cardapios": {
            "get": {
                "tags": [
                    "Categorias"
                ],
                "summary": "Mostra Categorias e cardápios relacionados",
                "description": "Show Categoria",
                "operationId": "b40ae9b12c1dbc229bedbd48c5fce4b4",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Parameter ",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "examples": {
                            "int": {
                                "summary": "An int value.",
                                "value": "1"
                            }
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": "OK"
                    }
                }
            }
        },
        "/api/categorias/{id}": {
            "put": {
                "tags": [
                    "Categorias"
                ],
                "summary": "Updates a Categoria",
                "operationId": "10e328fae70571e433bffe3e57bf6e51",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Update Categoria",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "examples": {
                            "int": {
                                "summary": "categoria_id",
                                "value": "1"
                            }
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nome": {
                                        "type": "string"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "nome": "Saída"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Categorias"
                ],
                "summary": "Delete Categoria",
                "description": "Delete Categoria",
                "operationId": "1babc7590157b13bebe0bfd683a029b2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Parameter ",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "examples": {
                            "int": {
                                "summary": "An int value.",
                                "value": "1"
                            }
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    }
}